+++ app/src/main/java/ru/netology/nmedia_practice/dto/Post.kt
@@ -5,9 +5,9 @@ data class Post (
     val published: String,
     val content: String,
     val likedByMe: Boolean = false,
-    val countLikes: Long = 0,
+    val countLikes: Int = 0,
     val send: Boolean = false,
-    val countSend: Long = 0,
-    val countView: Long = 0,
+    val countSend: Int = 0,
+    val countView: Int = 0,
     val videoUrl: String? = null
 )
